requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.4
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
pymongo==3.12.1
pytz==2022.7
sqlparse==0.2.4
